Moved admin commands from init.py to new file/cog admin.py
Created ProphetBot class as extension of discord.ext.commands.Bot
 - Mostly done to implement a custom on_command_error method
 - Should probably break this out into its own file

Renamed mgmt.py to bpdia.py for clarity

Added 'sheet' command to bpdia.py

Added 'all' as accepted parameter to 'reload'
